Move find_use stuff to it's own file
